[jupyter-gcs-contents-manager] Improve error message for creating a file in the virtual top-level directory
This change fixes a usability issue in the error message returned when a user of the CombinedContentsManager tries to save a file to the virtual, top-level directory.

Previously, that would result in a 404 error due to the contents manager not being able to find a nested contents manager for the top-level.

With this new change, it will return a 403 error saying the top-level is read-only

+++ jupyter-gcs-contents-manager/gcs_contents_manager.py
@@ -548,18 +548,17 @@ class CombinedContentsManager(ContentsManager):
       if path == path_prefix or path.startswith(path_prefix + '/'):
         relative_path = path[len(path_prefix):]
         return self._content_managers[path_prefix], relative_path, path_prefix
-    raise HTTPError(404, 'No content manager defined for "{}"'.format(path))
+    return None, '', path
 
   def is_hidden(self, path):
     try:
       cm, relative_path, unused_path_prefix = self._content_manager_for_path(
           path)
+      if not cm:
+        return False
       return cm.is_hidden(relative_path)
     except HTTPError as err:
-      if err.status_code == 404:
-        return False
-      else:
-        raise err
+      raise err
     except Exception as ex:
       raise HTTPError(
           500, 'Internal server error: [{}] {}'.format(type(ex), str(ex)))
@@ -568,12 +567,11 @@ class CombinedContentsManager(ContentsManager):
     try:
       cm, relative_path, unused_path_prefix = self._content_manager_for_path(
           path)
+      if not cm:
+        return False
       return cm.file_exists(relative_path)
     except HTTPError as err:
-      if err.status_code == 404:
-        return False
-      else:
-        raise err
+      raise err
     except Exception as ex:
       raise HTTPError(
           500, 'Internal server error: [{}] {}'.format(type(ex), str(ex)))
@@ -584,12 +582,11 @@ class CombinedContentsManager(ContentsManager):
     try:
       cm, relative_path, unused_path_prefix = self._content_manager_for_path(
           path)
+      if not cm:
+        return False
       return cm.dir_exists(relative_path)
     except HTTPError as err:
-      if err.status_code == 404:
-        return False
-      else:
-        raise err
+      raise err
     except Exception as ex:
       raise HTTPError(
           500, 'Internal server error: [{}] {}'.format(type(ex), str(ex)))
@@ -630,6 +627,8 @@ class CombinedContentsManager(ContentsManager):
       return dir_obj
     try:
       cm, relative_path, path_prefix = self._content_manager_for_path(path)
+      if not cm:
+        raise HTTPError(404, 'No content manager defined for "{}"'.format(path))
       model = cm.get(relative_path, content=content, type=type, format=format)
       self._make_model_relative(model, path_prefix)
       return model
@@ -648,6 +647,8 @@ class CombinedContentsManager(ContentsManager):
       cm, relative_path, path_prefix = self._content_manager_for_path(path)
       if relative_path in ['', '/']:
         raise HTTPError(403, 'The top-level directory contents are read-only')
+      if not cm:
+        raise HTTPError(404, 'No content manager defined for "{}"'.format(path))
 
       if 'path' in model:
         model['path'] = relative_path
@@ -669,6 +670,8 @@ class CombinedContentsManager(ContentsManager):
       cm, relative_path, path_prefix = self._content_manager_for_path(path)
       if relative_path in ['', '/']:
         raise HTTPError(403, 'The top-level directory contents are read-only')
+      if not cm:
+        raise HTTPError(404, 'No content manager defined for "{}"'.format(path))
       return cm.delete_file(relative_path)
     except OSError as err:
       # The built-in file contents manager will not attempt to wrap permissions
@@ -696,10 +699,14 @@ class CombinedContentsManager(ContentsManager):
       old_cm, old_relative_path, _ = self._content_manager_for_path(old_path)
       if old_relative_path in ['', '/']:
         raise HTTPError(403, 'The top-level directory contents are read-only')
+      if not old_cm:
+        raise HTTPError(404, 'No content manager defined for "{}"'.format(old_path))
 
       new_cm, new_relative_path, _ = self._content_manager_for_path(new_path)
       if new_relative_path in ['', '/']:
         raise HTTPError(403, 'The top-level directory contents are read-only')
+      if not new_cm:
+        raise HTTPError(404, 'No content manager defined for "{}"'.format(new_path))
 
       if old_cm != new_cm:
         raise HTTPError(400, 'Unsupported rename across file systems')
